+++ package.json
@@ -12,7 +12,7 @@
   "devDependencies": {
     "eslint": "7.13.0",
     "eslint-config-htmlacademy": "0.6.0",
-    "nanoid": "^3.1.20",
+    "nanoid": "3.1.20",
     "webpack": "4.44.2",
     "webpack-cli": "3.3.12",
     "webpack-dev-server": "3.11.0"
